[FE_Feat]사용자 접근권한에 따른 UI 표시 변경

+++ client/src/Components/Comments.js
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { useCookies } from 'react-cookie';
-import { useDispatch } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import styled from 'styled-components';
 import {
   deleteCommentThunk,
@@ -102,6 +102,7 @@ const Comments = ({
   setCommentsData,
   handleRender,
   render,
+  memberId,
 }) => {
   const dispatch = useDispatch();
 
@@ -111,37 +112,41 @@ const Comments = ({
   const [isEditOn, setIsEditOn] = useState(false);
 
   const addComment = () => {
-    async function postComment() {
-      const response = await dispatch(
-        postCommentThunk({
-          questionId: currentQuestion.questionId,
-          text: newComment,
-          cookie: cookies.access_token,
-        })
-      ).then((res) => {
-        setNewComment('');
-        handleRender(!render);
-      });
+    if (confirm('댓글을 추가하시겠습니까?')) {
+      async function postComment() {
+        const response = await dispatch(
+          postCommentThunk({
+            questionId: currentQuestion.questionId,
+            text: newComment,
+            cookie: cookies.access_token,
+          })
+        ).then((res) => {
+          setNewComment('');
+          handleRender(!render);
+        });
+      }
+      postComment();
     }
-    postComment();
   };
 
   const deleteComment = (event) => {
-    async function deleteComment() {
-      const response = await dispatch(
-        deleteCommentThunk({
-          questionId: currentQuestion.questionId,
-          commentId: commentsData[event.target.id].commentId,
-          cookie: cookies.access_token,
-        })
-      ).then((data) => handleRender(!render));
+    if (confirm('정말 삭제하시겠습니까?')) {
+      async function deleteComment() {
+        const response = await dispatch(
+          deleteCommentThunk({
+            questionId: currentQuestion.questionId,
+            commentId: commentsData[event.target.id].commentId,
+            cookie: cookies.access_token,
+          })
+        ).then((data) => handleRender(!render));
+      }
+      deleteComment();
+      setCommentsData(
+        commentsData.filter(
+          (a) => a.commentId !== commentsData[event.target.id].commentId
+        )
+      );
     }
-    deleteComment();
-    setCommentsData(
-      commentsData.filter(
-        (a) => a.commentId !== commentsData[event.target.id].commentId
-      )
-    );
   };
 
   const toggleEdit = (event) => {
@@ -151,21 +156,23 @@ const Comments = ({
   };
 
   const editComment = () => {
-    async function editComment() {
-      const response = await dispatch(
-        patchCommentThunk({
-          questionId: currentQuestion.questionId,
-          commentId: currentCommentId,
-          text: newComment,
-          cookie: cookies.access_token,
-        })
-      ).then((res) => {
-        setNewComment('');
-        handleRender(!render);
-      });
+    if (confirm('댓글을 수정하시겠습니까?')) {
+      async function editComment() {
+        const response = await dispatch(
+          patchCommentThunk({
+            questionId: currentQuestion.questionId,
+            commentId: currentCommentId,
+            text: newComment,
+            cookie: cookies.access_token,
+          })
+        ).then((res) => {
+          setNewComment('');
+          handleRender(!render);
+        });
+      }
+      editComment();
+      setIsEditOn(false);
     }
-    editComment();
-    setIsEditOn(false);
   };
 
   const cancelEdit = () => {
@@ -179,22 +186,24 @@ const Comments = ({
         <li key={`${currentQuestion.questionId}_${idx}`}>
           <div className="Comment_Block">
             <p>{item.text}</p>
-            <Comment_Buttons_Wrapper>
-              <button
-                className="Edit_Comment_Button"
-                onClick={toggleEdit}
-                id={idx}
-              >
-                수정
-              </button>
-              <button
-                className="Delete_Comment_Button"
-                onClick={deleteComment}
-                id={idx}
-              >
-                삭제
-              </button>
-            </Comment_Buttons_Wrapper>
+            {memberId === currentQuestion.memberId ? (
+              <Comment_Buttons_Wrapper>
+                <button
+                  className="Edit_Comment_Button"
+                  onClick={toggleEdit}
+                  id={idx}
+                >
+                  수정
+                </button>
+                <button
+                  className="Delete_Comment_Button"
+                  onClick={deleteComment}
+                  id={idx}
+                >
+                  삭제
+                </button>
+              </Comment_Buttons_Wrapper>
+            ) : null}
           </div>
           <Custom_Hr />
         </li>
